Added tab order calculations in core.js, modified test suites for nav, vertical components and core
