+++ tests/ui/components/navigation/navigationcomponent.js
@@ -1,7 +1,5 @@
 import DOM from '../../../../src/ui/dom/dom';
-import NavigationComponent, { Tab } from '../../../../src/ui/components/navigation/navigationcomponent';
-import VerticalPagesConfig from '../../../../src/core/models/verticalpagesconfig';
-import StorageKeys from '../../../../src/core/storage/storagekeys';
+import { Tab } from '../../../../src/ui/components/navigation/navigationcomponent';
 
 // The DOM doesn't exist within components in the JEST environment,
 // so we have to provide it to our DOM API properly.
@@ -102,138 +100,5 @@ describe('navigation component configuration', () => {
 
     const tab = Tab.from([tab1Config, tab2Config, tab3Config]);
     expect(tab).toMatchObject(expected);
-  });
-
-  const mockedCore = (config) => {
-    const verticalsConfig = new VerticalPagesConfig(config);
-    return {
-      globalStorage: {
-        getState: (storageKey) => {
-          expect(storageKey).toEqual(StorageKeys.VERTICAL_PAGES_CONFIG);
-          return verticalsConfig;
-        }
-      }
-    };
-  };
-
-  it('component supports default tab ordering from config', () => {
-    const tabConfig = [
-      {
-        label: 'This this is my second label',
-        url: '/tab1/',
-        verticalKey: 'tab1'
-      },
-      {
-        label: 'This is my label',
-        url: '/tab2/',
-        verticalKey: 'tab2',
-        isFirst: true
-      }
-    ];
-
-    // Finally construct our component
-    let navComponent = new NavigationComponent({
-      container: '.test-component'
-    }, {
-      core: mockedCore(tabConfig)
-    });
-
-    const defaultOrder = navComponent.getDefaultTabOrder(tabConfig);
-    expect(defaultOrder).toMatchObject(['tab2', 'tab1']);
-  });
-
-  it('component supports default tab ordering from params', () => {
-    const tabConfig = [
-      {
-        label: 'This this is my second label',
-        url: '/tab1/',
-        verticalKey: 'tab1'
-      },
-      {
-        label: 'This is my label',
-        url: '/tab2/',
-        verticalKey: 'tab2'
-      }
-    ];
-
-    // Finally construct our component
-    let navComponent = new NavigationComponent({
-      container: '.test-component'
-    }, {
-      core: mockedCore(tabConfig)
-    });
-
-    let params = new URLSearchParams('tabOrder=tab2,tab1');
-    const defaultOrder = navComponent.getDefaultTabOrder(tabConfig, params);
-    expect(defaultOrder).toMatchObject(['tab2', 'tab1']);
-  });
-
-  it('merging two tab orders', () => {
-    const tabConfig = [
-      {
-        label: 'This this is my second label',
-        url: '/tab1/',
-        verticalKey: 'tab1'
-      },
-      {
-        label: 'This is my label',
-        url: '/tab2/',
-        verticalKey: 'tab2'
-      }
-    ];
-
-    // Finally construct our component
-    let navComponent = new NavigationComponent({
-      container: '.test-component'
-    }, {
-      core: mockedCore(tabConfig)
-    });
-
-    const tabOrder1 = ['tab1', 'tab2'];
-    const tabOrder2 = ['tab2', 'tab3'];
-
-    const defaultOrder = navComponent.mergeTabOrder(tabOrder1, tabOrder2);
-    expect(defaultOrder).toMatchObject(['tab1', 'tab2', 'tab3']);
-  });
-
-  it('generates tab url', () => {
-    const tabConfig = [
-      {
-        label: 'This this is my second label',
-        url: '/tab1/',
-        verticalKey: 'tab1'
-      }
-    ];
-
-    // Finally construct our component
-    let navComponent = new NavigationComponent({
-      container: '.test-component'
-    }, {
-      core: mockedCore(tabConfig)
-    });
-
-    const params = new URLSearchParams('query=yes');
-    const url = navComponent.generateTabUrl(tabConfig[0].url, params);
-    expect(url).toEqual('/tab1/?query=yes&tabOrder=tab1');
-  });
-
-  it('generates tab url without params', () => {
-    const tabConfig = [
-      {
-        label: 'This this is my second label',
-        url: '/tab1/',
-        verticalKey: 'tab1'
-      }
-    ];
-
-    // Finally construct our component
-    let navComponent = new NavigationComponent({
-      container: '.test-component'
-    }, {
-      core: mockedCore(tabConfig)
-    });
-
-    const url = navComponent.generateTabUrl(tabConfig[0].url);
-    expect(url).toEqual('/tab1/?tabOrder=tab1');
   });
 });
