+++ tests/ui/components/results/verticalresultscomponent.js
@@ -18,7 +18,11 @@ const mockCore = {
   },
   getStaticFilterNodes: () => [],
   getFacetFilterNodes: () => [],
-  getLocationRadiusFilterNode: () => null
+  getLocationRadiusFilterNode: () => null,
+  getDefaultTabOrder: () => [],
+  mergeTabOrder: () => [],
+  getUrlParams: () => [],
+  generateTabUrl: () => 'defaultTabUrl'
 };
 
 DOM.setup(document, new DOMParser());
@@ -123,7 +127,7 @@ describe('vertical results component', () => {
     });
 
     it('if unset defaults to vertical key', () => {
-      expect(component.getVerticalURL()).toEqual('key.html?query=my-query&otherParam=123');
+      expect(component.getVerticalURL()).toEqual('defaultTabUrl');
     });
 
     it('if null defaults to vertical key', () => {
@@ -132,13 +136,13 @@ describe('vertical results component', () => {
       });
       component.query = 'my-query';
       component.verticalKey = 'key';
-      expect(component.getVerticalURL()).toEqual('key.html?query=my-query&otherParam=123');
+      expect(component.getVerticalURL()).toEqual('defaultTabUrl');
     });
 
     it('works with transformData', () => {
       expect(component.getVerticalURL({
         verticalURL: 'transform-data'
-      })).toEqual('transform-data?query=my-query&otherParam=123');
+      })).toEqual('defaultTabUrl');
     });
 
     it('defaults to matching config in verticalPages', () => {
@@ -146,7 +150,7 @@ describe('vertical results component', () => {
         verticalKey: 'key',
         url: 'vertical-pages'
       }];
-      expect(component.getVerticalURL()).toEqual('vertical-pages?query=my-query&otherParam=123');
+      expect(component.getVerticalURL()).toEqual('defaultTabUrl');
     });
 
     it('can be set', () => {
@@ -155,7 +159,7 @@ describe('vertical results component', () => {
       });
       component.query = 'my-query';
       component.verticalKey = 'key';
-      expect(component.getVerticalURL()).toEqual('vertical-url?query=my-query&otherParam=123');
+      expect(component.getVerticalURL()).toEqual('defaultTabUrl');
     });
   });
 });
